backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-magic==0.4.27
anthropic==0.34.2
yara-python==4.5.1
reportlab==4.2.2
pydantic==2.8.2
aiofiles==24.1.0
